Enhance Android Vulkan Renderer Integration
- Updated CMake configuration to enforce SDL usage and set default renderer to Vulkan for Android builds.
- Modified IdTech3Activity to implement SurfaceView for Vulkan rendering, enabling proper lifecycle management of the rendering surface.
- Introduced JNI methods for surface lifecycle callbacks, allowing the engine to manage Vulkan swapchain creation and destruction.
- Improved thread safety in Android platform management by adding mutexes for window handling.
- Enhanced Vulkan initialization with additional debug logging and error handling for better traceability during setup.

+++ src/client/cl_main.c
@@ -26,6 +26,7 @@ Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
 #include "../sdl/sdl_glw.h"
 #endif
 #include <limits.h>
+#include <stdint.h>
 #ifdef USE_CURL
 #include "cl_net_enhanced.h"
 #endif
@@ -3786,7 +3787,8 @@ fprintf(stderr, "About to enter renderer loading logic\n");
 	{
 		extern refimport_t ri;
 		ri = rimp;
-		fprintf(stderr, "DEBUG: CL_InitRef set global ri.Printf to %p\n", (void*)ri.Printf);
+		fprintf(stderr, "DEBUG: CL_InitRef set global ri.Printf to %p\n",
+		        (void*)(uintptr_t)ri.Printf);
 	}
 	
 	// Validate that all Vulkan function pointers were properly assigned
